polish

+++ test/src/main.rs
@@ -61,7 +61,7 @@ mod tests {
         let init_script = "create_table_1.sql".to_string();
         conn.init(init_script).await?;
         let insert_id = conn.insert(user.clone()).run().await?;
-        let updated_rows: i32 = conn.query_update("insert into user (id) values (3)".to_string()).run().await?;
+        let updated_rows: usize = conn.query_update("insert into user (id) values (3)".to_string()).run().await?;
 
         let query = format!("select * from user where name like {}", conn.protect("%oh%".to_string()));
         let result_set: Vec<Row> = conn.query(query).run().await?;
